sciter input & conn_type and other InvokeUi impl
