+++ src/ui/remote.rs
@@ -22,14 +22,14 @@ use clipboard::{
     cliprdr::CliprdrClientContext, create_cliprdr_context as create_clipboard_file_context,
     get_rx_clip_client, server_clip_file,
 };
-use enigo::{self};
-use hbb_common::{allow_err, log, message_proto::*};
+
+use hbb_common::{allow_err, log, message_proto::*, rendezvous_proto::ConnType};
 
 #[cfg(windows)]
 use crate::clipboard_file::*;
 use crate::{
     client::*,
-    ui_session_interface::{InvokeUi, Session},
+    ui_session_interface::{InvokeUi, Session, IS_IN},
 };
 
 type Video = AssetPtr<video_destination>;
@@ -37,9 +37,6 @@ type Video = AssetPtr<video_destination>;
 lazy_static::lazy_static! {
     static ref VIDEO: Arc<Mutex<Option<Video>>> = Default::default();
 }
-
-static IS_IN: AtomicBool = AtomicBool::new(false);
-static KEYBOARD_HOOKED: AtomicBool = AtomicBool::new(false);
 
 #[cfg(windows)]
 static mut IS_ALT_GR: bool = false;
@@ -397,255 +394,28 @@ impl sciter::EventHandler for SciterSession {
 impl SciterSession {
     pub fn new(cmd: String, id: String, password: String, args: Vec<String>) -> Self {
         let session: Session<SciterHandler> = Session {
-            cmd,
+            cmd: cmd.clone(),
             id: id.clone(),
             password: password.clone(),
             args,
             ..Default::default()
         };
-        session.lc.write().unwrap().initialize(
-            id,
-            session.is_file_transfer(),
-            session.is_port_forward(),
-        );
+
+        let conn_type = if cmd.eq("--file-transfer") {
+            ConnType::FILE_TRANSFER
+        } else if cmd.eq("--port-forward") {
+            ConnType::PORT_FORWARD
+        } else if cmd.eq("--rdp") {
+            ConnType::RDP
+        } else {
+            ConnType::DEFAULT_CONN
+        };
+
+        session.lc.write().unwrap().initialize(id, conn_type);
 
         Self(session)
     }
 
-    // TODO
-    fn start_keyboard_hook(&'static self) {
-        if self.is_port_forward() || self.is_file_transfer() {
-            return;
-        }
-        if KEYBOARD_HOOKED.swap(true, Ordering::SeqCst) {
-            return;
-        }
-        log::info!("keyboard hooked");
-        let me = self.clone();
-        let peer = self.peer_platform();
-        let is_win = peer == "Windows";
-        #[cfg(windows)]
-        crate::platform::windows::enable_lowlevel_keyboard(std::ptr::null_mut() as _);
-        std::thread::spawn(move || {
-            // This will block.
-            std::env::set_var("KEYBOARD_ONLY", "y"); // pass to rdev
-            use rdev::{EventType::*, *};
-            let func = move |evt: Event| {
-                if !IS_IN.load(Ordering::SeqCst) || !SERVER_KEYBOARD_ENABLED.load(Ordering::SeqCst)
-                {
-                    return;
-                }
-                let (key, down) = match evt.event_type {
-                    KeyPress(k) => (k, 1),
-                    KeyRelease(k) => (k, 0),
-                    _ => return,
-                };
-                let alt = get_key_state(enigo::Key::Alt);
-                #[cfg(windows)]
-                let ctrl = {
-                    let mut tmp = get_key_state(enigo::Key::Control);
-                    unsafe {
-                        if IS_ALT_GR {
-                            if alt || key == Key::AltGr {
-                                if tmp {
-                                    tmp = false;
-                                }
-                            } else {
-                                IS_ALT_GR = false;
-                            }
-                        }
-                    }
-                    tmp
-                };
-                #[cfg(not(windows))]
-                let ctrl = get_key_state(enigo::Key::Control);
-                let shift = get_key_state(enigo::Key::Shift);
-                #[cfg(windows)]
-                let command = crate::platform::windows::get_win_key_state();
-                #[cfg(not(windows))]
-                let command = get_key_state(enigo::Key::Meta);
-                let control_key = match key {
-                    Key::Alt => Some(ControlKey::Alt),
-                    Key::AltGr => Some(ControlKey::RAlt),
-                    Key::Backspace => Some(ControlKey::Backspace),
-                    Key::ControlLeft => {
-                        // when pressing AltGr, an extra VK_LCONTROL with a special
-                        // scancode with bit 9 set is sent, let's ignore this.
-                        #[cfg(windows)]
-                        if evt.scan_code & 0x200 != 0 {
-                            unsafe {
-                                IS_ALT_GR = true;
-                            }
-                            return;
-                        }
-                        Some(ControlKey::Control)
-                    }
-                    Key::ControlRight => Some(ControlKey::RControl),
-                    Key::DownArrow => Some(ControlKey::DownArrow),
-                    Key::Escape => Some(ControlKey::Escape),
-                    Key::F1 => Some(ControlKey::F1),
-                    Key::F10 => Some(ControlKey::F10),
-                    Key::F11 => Some(ControlKey::F11),
-                    Key::F12 => Some(ControlKey::F12),
-                    Key::F2 => Some(ControlKey::F2),
-                    Key::F3 => Some(ControlKey::F3),
-                    Key::F4 => Some(ControlKey::F4),
-                    Key::F5 => Some(ControlKey::F5),
-                    Key::F6 => Some(ControlKey::F6),
-                    Key::F7 => Some(ControlKey::F7),
-                    Key::F8 => Some(ControlKey::F8),
-                    Key::F9 => Some(ControlKey::F9),
-                    Key::LeftArrow => Some(ControlKey::LeftArrow),
-                    Key::MetaLeft => Some(ControlKey::Meta),
-                    Key::MetaRight => Some(ControlKey::RWin),
-                    Key::Return => Some(ControlKey::Return),
-                    Key::RightArrow => Some(ControlKey::RightArrow),
-                    Key::ShiftLeft => Some(ControlKey::Shift),
-                    Key::ShiftRight => Some(ControlKey::RShift),
-                    Key::Space => Some(ControlKey::Space),
-                    Key::Tab => Some(ControlKey::Tab),
-                    Key::UpArrow => Some(ControlKey::UpArrow),
-                    Key::Delete => {
-                        if is_win && ctrl && alt {
-                            // me.ctrl_alt_del(); // TODO
-                            return;
-                        }
-                        Some(ControlKey::Delete)
-                    }
-                    Key::Apps => Some(ControlKey::Apps),
-                    Key::Cancel => Some(ControlKey::Cancel),
-                    Key::Clear => Some(ControlKey::Clear),
-                    Key::Kana => Some(ControlKey::Kana),
-                    Key::Hangul => Some(ControlKey::Hangul),
-                    Key::Junja => Some(ControlKey::Junja),
-                    Key::Final => Some(ControlKey::Final),
-                    Key::Hanja => Some(ControlKey::Hanja),
-                    Key::Hanji => Some(ControlKey::Hanja),
-                    Key::Convert => Some(ControlKey::Convert),
-                    Key::Print => Some(ControlKey::Print),
-                    Key::Select => Some(ControlKey::Select),
-                    Key::Execute => Some(ControlKey::Execute),
-                    Key::PrintScreen => Some(ControlKey::Snapshot),
-                    Key::Help => Some(ControlKey::Help),
-                    Key::Sleep => Some(ControlKey::Sleep),
-                    Key::Separator => Some(ControlKey::Separator),
-                    Key::KpReturn => Some(ControlKey::NumpadEnter),
-                    Key::Kp0 => Some(ControlKey::Numpad0),
-                    Key::Kp1 => Some(ControlKey::Numpad1),
-                    Key::Kp2 => Some(ControlKey::Numpad2),
-                    Key::Kp3 => Some(ControlKey::Numpad3),
-                    Key::Kp4 => Some(ControlKey::Numpad4),
-                    Key::Kp5 => Some(ControlKey::Numpad5),
-                    Key::Kp6 => Some(ControlKey::Numpad6),
-                    Key::Kp7 => Some(ControlKey::Numpad7),
-                    Key::Kp8 => Some(ControlKey::Numpad8),
-                    Key::Kp9 => Some(ControlKey::Numpad9),
-                    Key::KpDivide => Some(ControlKey::Divide),
-                    Key::KpMultiply => Some(ControlKey::Multiply),
-                    Key::KpDecimal => Some(ControlKey::Decimal),
-                    Key::KpMinus => Some(ControlKey::Subtract),
-                    Key::KpPlus => Some(ControlKey::Add),
-                    Key::CapsLock | Key::NumLock | Key::ScrollLock => {
-                        return;
-                    }
-                    Key::Home => Some(ControlKey::Home),
-                    Key::End => Some(ControlKey::End),
-                    Key::Insert => Some(ControlKey::Insert),
-                    Key::PageUp => Some(ControlKey::PageUp),
-                    Key::PageDown => Some(ControlKey::PageDown),
-                    Key::Pause => Some(ControlKey::Pause),
-                    _ => None,
-                };
-                let mut key_event = KeyEvent::new();
-                if let Some(k) = control_key {
-                    key_event.set_control_key(k);
-                } else {
-                    let mut chr = match evt.name {
-                        Some(ref s) => {
-                            if s.len() <= 2 {
-                                // exclude chinese characters
-                                s.chars().next().unwrap_or('\0')
-                            } else {
-                                '\0'
-                            }
-                        }
-                        _ => '\0',
-                    };
-                    if chr == '·' {
-                        // special for Chinese
-                        chr = '`';
-                    }
-                    if chr == '\0' {
-                        chr = match key {
-                            Key::Num1 => '1',
-                            Key::Num2 => '2',
-                            Key::Num3 => '3',
-                            Key::Num4 => '4',
-                            Key::Num5 => '5',
-                            Key::Num6 => '6',
-                            Key::Num7 => '7',
-                            Key::Num8 => '8',
-                            Key::Num9 => '9',
-                            Key::Num0 => '0',
-                            Key::KeyA => 'a',
-                            Key::KeyB => 'b',
-                            Key::KeyC => 'c',
-                            Key::KeyD => 'd',
-                            Key::KeyE => 'e',
-                            Key::KeyF => 'f',
-                            Key::KeyG => 'g',
-                            Key::KeyH => 'h',
-                            Key::KeyI => 'i',
-                            Key::KeyJ => 'j',
-                            Key::KeyK => 'k',
-                            Key::KeyL => 'l',
-                            Key::KeyM => 'm',
-                            Key::KeyN => 'n',
-                            Key::KeyO => 'o',
-                            Key::KeyP => 'p',
-                            Key::KeyQ => 'q',
-                            Key::KeyR => 'r',
-                            Key::KeyS => 's',
-                            Key::KeyT => 't',
-                            Key::KeyU => 'u',
-                            Key::KeyV => 'v',
-                            Key::KeyW => 'w',
-                            Key::KeyX => 'x',
-                            Key::KeyY => 'y',
-                            Key::KeyZ => 'z',
-                            Key::Comma => ',',
-                            Key::Dot => '.',
-                            Key::SemiColon => ';',
-                            Key::Quote => '\'',
-                            Key::LeftBracket => '[',
-                            Key::RightBracket => ']',
-                            Key::BackSlash => '\\',
-                            Key::Minus => '-',
-                            Key::Equal => '=',
-                            Key::BackQuote => '`',
-                            _ => '\0',
-                        }
-                    }
-                    if chr != '\0' {
-                        if chr == 'l' && is_win && command {
-                            // me.lock_screen(); // TODO
-                            return;
-                        }
-                        key_event.set_chr(chr as _);
-                    } else {
-                        log::error!("Unknown key {:?}", evt);
-                        return;
-                    }
-                }
-                // me.key_down_or_up(down, key_event, alt, ctrl, shift, command); // TODO
-            };
-            if let Err(error) = rdev::listen(func) {
-                log::error!("rdev: {:?}", error);
-            }
-        });
-    }
-
-    // TODO
     fn get_custom_image_quality(&mut self) -> Value {
         let mut v = Value::array(0);
         for x in self.lc.read().unwrap().custom_image_quality.iter() {
@@ -654,7 +424,6 @@ impl SciterSession {
         v
     }
 
-    // TODO
     fn supported_hwcodec(&self) -> Value {
         #[cfg(feature = "hwcodec")]
         {
@@ -679,7 +448,6 @@ impl SciterSession {
         }
     }
 
-    // TODO
     fn save_size(&mut self, x: i32, y: i32, w: i32, h: i32) {
         let size = (x, y, w, h);
         let mut config = self.load_config();
@@ -765,7 +533,6 @@ impl SciterSession {
         pi
     }
 
-    // close_state sciter only
     fn save_close_state(&mut self, k: String, v: String) {
         self.close_state.insert(k, v);
     }
